Merge develop for release 1.14.0

+++ src/main/java/org/isf/lab/service/LabIoOperations.java
@@ -1,6 +1,6 @@
 /*
  * Open Hospital (www.open-hospital.org)
- * Copyright © 2006-2023 Informatici Senza Frontiere (info@informaticisenzafrontiere.org)
+ * Copyright © 2006-2024 Informatici Senza Frontiere (info@informaticisenzafrontiere.org)
  *
  * Open Hospital is a free and open source software for healthcare data management.
  *
@@ -27,12 +27,14 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
+import org.isf.exa.model.Exam;
 import org.isf.lab.model.Laboratory;
 import org.isf.lab.model.LaboratoryForPrint;
 import org.isf.lab.model.LaboratoryRow;
 import org.isf.patient.model.Patient;
 import org.isf.utils.db.TranslateOHServiceException;
 import org.isf.utils.exception.OHServiceException;
+import org.isf.utils.exception.model.OHExceptionMessage;
 import org.isf.utils.pagination.PageInfo;
 import org.isf.utils.pagination.PagedResponse;
 import org.isf.utils.time.TimeTools;
@@ -43,23 +45,6 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-/**
- * ------------------------------------------
- * lab.service.LabIoOperations - laboratory exam database io operations
- * -----------------------------------------
- * modification history
- * 02/03/2006 - theo - first beta version
- * 10/11/2006 - ross - added editing capability.
- * 					   new fields data esame, sex, age, material, inout flag added
- * 21/06/2008 - ross - do not add 1 to toDate!.
- *                     the selection date switched to exam date,
- * 04/01/2009 - ross - do not use roll, use add(week,-1)!
- *                     roll does not change the year!
- * 16/11/2012 - mwithi - added logging capability
- * 					   - to do lock management
- * 04/02/2013 - mwithi - lock management done
- * ------------------------------------------
- */
 @Service
 @Transactional(rollbackFor = OHServiceException.class)
 @TranslateOHServiceException
@@ -72,9 +57,10 @@ public class LabIoOperations {
 	private LabRowIoOperationRepository rowRepository;
 
 	/**
-	 * Return a list of results ({@link LaboratoryRow}s) for passed lab entry.
+	 * Return a list of results ({@link LaboratoryRow}s) for passed lab code.
+	 *
 	 * @param code - the {@link Laboratory} record ID.
-	 * @return the list of {@link LaboratoryRow}s. It could be <code>empty</code>
+	 * @return the list of {@link LaboratoryRow}s. It could be {@code empty}
 	 * @throws OHServiceException
 	 */
 	public List<LaboratoryRow> getLabRow(Integer code) throws OHServiceException {
@@ -84,15 +70,15 @@ public class LabIoOperations {
 	/**
 	 * Return the list of exams ({@link Laboratory}s) divided by pages.
 	 * 
-	 * @param onWeek
+	 * @param oneWeek
 	 * @param pageNo
 	 * @param pageSize
 	 * @return the list of {@link Laboratory}s (could be empty)
 	 * @throws OHServiceException
 	 */
-	public List<Laboratory> getLaboratory(boolean onWeek, int pageNo, int pageSize) throws OHServiceException {
+	public List<Laboratory> getLaboratory(boolean oneWeek, int pageNo, int pageSize) throws OHServiceException {
 		Pageable pageable = PageRequest.of(pageNo, pageSize);
-		if (onWeek) {
+		if (oneWeek) {
 			LocalDateTime time2 = TimeTools.getDateToday24();
 			LocalDateTime time1 = time2.minusWeeks(1);
 			return repository.findByLabDateBetweenOrderByLabDateDesc(time1, time2, pageable).getContent();
@@ -100,7 +86,7 @@ public class LabIoOperations {
 		}
 		return repository.findAll(pageable).getContent();
 	}
-	
+
 	public PagedResponse<Laboratory> getLaboratoryPageable(boolean oneWeek, int pageNo, int pageSize) throws OHServiceException {
 		Pageable pageable = PageRequest.of(pageNo, pageSize);
 		if (oneWeek) {
@@ -112,9 +98,10 @@ public class LabIoOperations {
 		Page<Laboratory> pagedResult = repository.findAll(pageable);
 		return setPaginationData(pagedResult);
 	}
-	
+
 	/**
-	 * Return the whole list of exams ({@link Laboratory}s) within last week.
+	 * Return the whole list of exams ({@link Laboratory}s) within the last week.
+	 *
 	 * @return the list of {@link Laboratory}s 
 	 * @throws OHServiceException
 	 */
@@ -125,8 +112,9 @@ public class LabIoOperations {
 	}
 
 	/**
-	 * Return a list of exams ({@link Laboratory}s) between specified dates and matching passed exam name
-	 * @param exam - the exam name as <code>String</code>
+	 * Return a list of exams ({@link Laboratory}s) between specified dates and matching passed exam name.
+	 *
+	 * @param exam - the exam name as {@code String}
 	 * @param dateFrom - the lower date for the range
 	 * @param dateTo - the highest date for the range
 	 * @return the list of {@link Laboratory}s 
@@ -141,8 +129,9 @@ public class LabIoOperations {
 	}
 
 	/**
-	 * Return a list of exams ({@link Laboratory}s) between specified dates and matching passed exam name
-	 * @param exam - the exam name as <code>String</code>
+	 * Return a list of exams ({@link Laboratory}s) between specified dates and matching passed exam name.
+	 *
+	 * @param exam - the exam name as {@code String}
 	 * @param dateFrom - the lower date for the range
 	 * @param dateTo - the highest date for the range
 	 * @param patient - the object of patient 
@@ -152,16 +141,16 @@ public class LabIoOperations {
 	public List<Laboratory> getLaboratory(String exam, LocalDateTime dateFrom, LocalDateTime dateTo, Patient patient) throws OHServiceException {
 		List<Laboratory> laboritories = new ArrayList<>();
 
-		if (!exam.equals("") && patient != null) {
+		if (!exam.isEmpty() && patient != null) {
 			laboritories = repository.findByLabDateBetweenAndExamDescriptionAndPatientCode(dateFrom, dateTo, exam, patient.getCode());
 		}
-		if (!exam.equals("") && patient == null) {
+		if (!exam.isEmpty() && patient == null) {
 			laboritories = repository.findByLabDateBetweenAndExam_DescriptionOrderByLabDateDesc(dateFrom, dateTo, exam);
 		}
-		if (patient != null && exam.equals("")) {
+		if (patient != null && exam.isEmpty()) {
 			laboritories = repository.findByLabDateBetweenAndPatientCode(dateFrom, dateTo, patient.getCode());
 		}
-		if (patient == null && exam.equals("")) {
+		if (patient == null && exam.isEmpty()) {
 			laboritories = repository.findByLabDateBetweenOrderByLabDateDesc(dateFrom, dateTo);
 		}
 		return laboritories;
@@ -169,6 +158,7 @@ public class LabIoOperations {
 
 	/**
 	 * Return a list of exams ({@link Laboratory}s) related to a {@link Patient}.
+	 *
 	 * @param aPatient - the {@link Patient}.
 	 * @return the list of {@link Laboratory}s related to the {@link Patient}.
 	 * @throws OHServiceException
@@ -178,8 +168,8 @@ public class LabIoOperations {
 	}
 
 	/**
-	 * Return a list of exams suitable for printing ({@link LaboratoryForPrint}s) 
-	 * within last year
+	 * Return a list of exams suitable for printing ({@link LaboratoryForPrint}s) within the last week.
+	 *
 	 * @return the list of {@link LaboratoryForPrint}s 
 	 * @throws OHServiceException
 	 */
@@ -191,8 +181,9 @@ public class LabIoOperations {
 
 	/**
 	 * Return a list of exams suitable for printing ({@link LaboratoryForPrint}s) 
-	 * between specified dates and matching passed exam name
-	 * @param exam - the exam name as <code>String</code>
+	 * between specified dates and matching passed exam name.
+	 *
+	 * @param exam - the exam name as {@code String}
 	 * @param dateFrom - the lower date for the range
 	 * @param dateTo - the highest date for the range
 	 * @return the list of {@link LaboratoryForPrint}s 
@@ -228,24 +219,25 @@ public class LabIoOperations {
 	}
 
 	/**
-	 * Insert a Laboratory exam {@link Laboratory} and return generated key. No commit is performed.
+	 * Insert a new Laboratory exam {@link Laboratory} and return the generated laboratory code.
+	 *
 	 * @param laboratory - the {@link Laboratory} to insert
-	 * @return the generated key
+	 * @return the newly persisted {@link Laboratory} object.
 	 * @throws OHServiceException
 	 */
-	private Integer newLaboratory(Laboratory laboratory) throws OHServiceException {
-		Laboratory savedLaboratory = repository.save(laboratory);
-		return savedLaboratory.getCode();
+	private Laboratory newLaboratory(Laboratory laboratory) throws OHServiceException {
+		return repository.save(laboratory);
 	}
 
 	/**
 	 * Inserts one Laboratory exam {@link Laboratory} (Procedure One)
+	 *
 	 * @param laboratory - the {@link Laboratory} to insert
-	 * @return <code>true</code> if the exam has been inserted, <code>false</code> otherwise
+	 * @return the newly persisted {@link Laboratory} object.
 	 * @throws OHServiceException
 	 */
-	public boolean newLabFirstProcedure(Laboratory laboratory) throws OHServiceException {
-		return newLaboratory(laboratory) > 0;
+	public Laboratory newLabFirstProcedure(Laboratory laboratory) throws OHServiceException {
+		return newLaboratory(laboratory);
 	}
 
 	/**
@@ -253,43 +245,29 @@ public class LabIoOperations {
 	 *
 	 * @param laboratory - the {@link Laboratory} to insert
 	 * @param labRow - the list of results ({@link String}s)
-	 * @return <code>true</code> if the exam has been inserted with all its results, <code>false</code> otherwise
+	 * @return the newly persisted {@link Laboratory} object.
 	 * @throws OHServiceException
 	 */
-	public boolean newLabSecondProcedure(Laboratory laboratory, List<String> labRow) throws OHServiceException {
-		boolean result = true;
-
-		int newCode = newLaboratory(laboratory);
-		if (newCode > 0) {
+	public Laboratory newLabSecondProcedure(Laboratory laboratory, List<String> labRow) throws OHServiceException {
+		Laboratory newLaboratory = newLaboratory(laboratory);
+		if (newLaboratory.getCode() > 0) {
 			for (String aLabRow : labRow) {
 				LaboratoryRow laboratoryRow = new LaboratoryRow();
 				laboratoryRow.setLabId(laboratory);
 				laboratoryRow.setDescription(aLabRow);
-
-				LaboratoryRow savedLaboratoryRow = rowRepository.save(laboratoryRow);
-				result = result && (savedLaboratoryRow != null);
+				rowRepository.save(laboratoryRow);
 			}
 		}
-		return result;
-	}
-	/**
-	 * Return a list of exams suitable for printing ({@link LaboratoryForPrint}s) 
-	 * within last year
-	 * @return the list of {@link LaboratoryForPrint}s 
-	 * @throws OHServiceException
-	 */
-	public List<LaboratoryForPrint> getLaboratoryForPrints() throws OHServiceException {
-		LocalDateTime time2 = TimeTools.getNow();
-		LocalDateTime time1 = time2.minusWeeks(1);
-		return getLaboratoryForPrint(null, time1, time2);
+		return newLaboratory;
 	}
 
 	/**
 	 * Return a list of exams suitable for printing ({@link LaboratoryForPrint}s) 
-	 * between specified dates and matching passed exam name
-	 * @param exam - the exam name as <code>String</code>
-	 * @param dateFrom - the lower date for the range
-	 * @param dateTo - the highest date for the range
+	 * between specified dates and matching passed exam name.
+	 *
+	 * @param exam - the exam name as {@code String}
+	 * @param dateFrom - the starting date for the date range
+	 * @param dateTo - the ending date for the date range
 	 * @return the list of {@link LaboratoryForPrint}s 
 	 * @throws OHServiceException
 	 */
@@ -317,89 +295,90 @@ public class LabIoOperations {
 	 *
 	 * @param laboratory - the {@link Laboratory} to insert
 	 * @param labRow - the list of results ({@link String}s)
-	 * @return <code>true</code> if the exam has been inserted with all its results, <code>false</code> otherwise
+	 * @return the newly persisted {@link Laboratory} object.
 	 * @throws OHServiceException
 	 */
-	public boolean newLabSecondProcedure2(Laboratory laboratory, List<LaboratoryRow> labRow) throws OHServiceException {
-		boolean result = true;
-
-		int newCode = newLaboratory(laboratory);
+	public Laboratory newLabSecondProcedure2(Laboratory laboratory, List<LaboratoryRow> labRow) throws OHServiceException {
+		Laboratory newLaboratory = newLaboratory(laboratory);
+		int newCode = newLaboratory.getCode();
 		if (newCode > 0) {
 			laboratory = repository.findById(newCode).orElse(null);
+			if (laboratory == null) {
+				throw new OHServiceException(new OHExceptionMessage("Laboratory with code '" + newCode + "' not found"));
+			}
 			for (LaboratoryRow aLabRow : labRow) {
 				aLabRow.setLabId(laboratory);
-				LaboratoryRow savedLaboratoryRow = rowRepository.save(aLabRow);
-				result = result && (savedLaboratoryRow != null);
+				rowRepository.save(aLabRow);
 			}
 		}
-		return result;
+		return newLaboratory;
 	}
 
 	/**
-	 * Update an already existing Laboratory exam {@link Laboratory}. No commit is performed.
+	 * Update an already existing Laboratory exam {@link Laboratory}.
+	 *
 	 * @param laboratory - the {@link Laboratory} to update
-	 * @return <code>true</code> if the exam has been updated with all its results, <code>false</code> otherwise
+	 * @return the updated {@link Laboratory} object.
 	 * @throws OHServiceException
 	 */
-	private boolean updateLaboratory(Laboratory laboratory) throws OHServiceException {
-		Laboratory savedLaboratory = repository.save(laboratory);
-		return savedLaboratory != null;
+	private Laboratory updateLaboratory(Laboratory laboratory) throws OHServiceException {
+		return repository.save(laboratory);
 	}
 
 	/**
 	 * Update an already existing Laboratory exam {@link Laboratory} (Procedure One).
-	 * If old exam was Procedure Two all its releated result are deleted.
+	 * If the old exam was Procedure Two then all its related results are deleted.
+	 *
 	 * @param laboratory - the {@link Laboratory} to update
-	 * @return <code>true</code> if the exam has been updated, <code>false</code> otherwise
+	 * @return {@code true} if the exam has been updated, {@code false} otherwise
 	 * @throws OHServiceException
 	 */
-	public boolean updateLabFirstProcedure(Laboratory laboratory) throws OHServiceException {
-		boolean result = updateLaboratory(laboratory);
-		rowRepository.deleteByLaboratory_Code(laboratory.getCode());
-		return result;
+	public Laboratory updateLabFirstProcedure(Laboratory laboratory) throws OHServiceException {
+		Laboratory updatedLaborator = updateLaboratory(laboratory);
+		rowRepository.deleteByLaboratory_Code(updatedLaborator.getCode());
+		return updatedLaborator;
 	}
 
 	/**
 	 * Update an already existing Laboratory exam {@link Laboratory} (Procedure Two).
 	 * Previous results are deleted and replaced with new ones.
 	 * @param laboratory - the {@link Laboratory} to update
-	 * @return <code>true</code> if the exam has been updated with all its results, <code>false</code> otherwise
+	 * @return the updated {@link Laboratory} object.
 	 * @throws OHServiceException
 	 */
-	public boolean updateLabSecondProcedure(Laboratory laboratory, List<String> labRow) throws OHServiceException {
-		boolean result = updateLabFirstProcedure(laboratory);
-		if (result) {
-			for (String aLabRow : labRow) {
-				LaboratoryRow laboratoryRow = new LaboratoryRow();
-				laboratoryRow.setLabId(laboratory);
-				laboratoryRow.setDescription(aLabRow);
-				rowRepository.save(laboratoryRow);
-			}
+	public Laboratory updateLabSecondProcedure(Laboratory laboratory, List<String> labRow) throws OHServiceException {
+		Laboratory updatedLaboratory = updateLabFirstProcedure(laboratory);
+		for (String aLabRow : labRow) {
+			LaboratoryRow laboratoryRow = new LaboratoryRow();
+			laboratoryRow.setLabId(laboratory);
+			laboratoryRow.setDescription(aLabRow);
+			rowRepository.save(laboratoryRow);
 		}
-		return result;
+		return updatedLaboratory;
 	}
 
 	/**
 	 * Delete a Laboratory exam {@link Laboratory} (Procedure One or Two).
 	 * Previous results, if any, are deleted as well.
 	 * @param aLaboratory - the {@link Laboratory} to delete
-	 * @return <code>true</code> if the exam has been deleted with all its results, if any. <code>false</code> otherwise
 	 * @throws OHServiceException
 	 */
-	public boolean deleteLaboratory(Laboratory aLaboratory) throws OHServiceException {
+	public void deleteLaboratory(Laboratory aLaboratory) throws OHServiceException {
 		Laboratory objToRemove = repository.findById(aLaboratory.getCode()).orElse(null);
+		if (objToRemove == null) {
+			throw new OHServiceException(new OHExceptionMessage("Laboratory object not found to delete"));
+		}
 		if (objToRemove.getExam().getProcedure() == 2) {
 			rowRepository.deleteByLaboratory_Code(objToRemove.getCode());
 		}
 		repository.deleteById(objToRemove.getCode());
-		return true;
 	}
 
 	/**
 	 * Checks if the code is already in use
 	 *
 	 * @param code - the laboratory code
-	 * @return <code>true</code> if the code is already in use, <code>false</code> otherwise
+	 * @return {@code true} if the code is already in use, {@code false} otherwise
 	 * @throws OHServiceException 
 	 */
 	public boolean isCodePresent(Integer code) throws OHServiceException {
@@ -409,18 +388,41 @@ public class LabIoOperations {
 	public Optional<Laboratory> getLaboratory(int code) throws OHServiceException {
 		return repository.findById(code);
 	}
-	
-	public PagedResponse<Laboratory> setPaginationData(Page<Laboratory> pages){
-		PagedResponse<Laboratory> data = new PagedResponse<Laboratory>();
+
+	public PagedResponse<Laboratory> getLaboratoryPageable(Exam exam, LocalDateTime dateFrom, LocalDateTime dateTo, Patient patient, int page, int size)
+					throws OHServiceException {
+		Page<Laboratory> laboritories = null;
+
+		if (exam != null && patient != null) {
+			laboritories = repository.findByLabDateBetweenAndExamDescriptionAndPatientCodePage(dateFrom, dateTo, exam, patient, PageRequest.of(page, size));
+		}
+		if (exam != null && patient == null) {
+			laboritories = repository.findByLabDateBetweenAndExam_DescriptionOrderByLabDateDescPage(dateFrom, dateTo, exam, PageRequest.of(page, size));
+		}
+		if (patient != null && exam == null) {
+			laboritories = repository.findByLabDateBetweenAndPatientCodePage(dateFrom, dateTo, patient, PageRequest.of(page, size));
+		}
+		if (patient == null && exam == null) {
+			laboritories = repository.findByLabDateBetweenOrderByLabDateDescPage(dateFrom, dateTo, PageRequest.of(page, size));
+		}
+		return setPaginationData(laboritories);
+	}
+
+	PagedResponse<Laboratory> setPaginationData(Page<Laboratory> pages) {
+		PagedResponse<Laboratory> data = new PagedResponse<>();
 		data.setData(pages.getContent());
-		PageInfo pageInfo = new PageInfo();
-		pageInfo.setSize(pages.getPageable().getPageSize());
-		pageInfo.setPage(pages.getPageable().getPageNumber());
-		pageInfo.setNbOfElements(pages.getNumberOfElements());
-		pageInfo.setTotalCount(pages.getTotalElements());
-		pageInfo.setHasPreviousPage(pages.hasPrevious());
-		pageInfo.setHasNextPage(pages.hasNext());
-		data.setPageInfo(pageInfo);
+		data.setPageInfo(PageInfo.from(pages));
 		return data;
 	}
+
+	/**
+	 * Count active {@link Laboratory}s
+	 * 
+	 * @return the number of recorded {@link Laboratory}s
+	 * @throws OHServiceException
+	 */
+	public long countAllActiveLabs() {
+		return this.repository.countAllActiveLabs();
+	}
+
 }
